requirements.txt
discord.py>=2.3.2
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
aiosqlite>=0.19.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
httpx>=0.25.2
jinja2>=3.1.2
aiofiles>=23.2.1
cryptography>=41.0.7
python-multipart>=0.0.6